+++ app/src/views/widgets/indicators/horizontal_chart_indicator/horizontal_chart_indicator.cpp
@@ -72,11 +72,11 @@ lv_obj_t* HorizontalChartIndicator::Create(lv_obj_t* parent, int32_t range_start
     return lv_chart_;
 }
 
-void HorizontalChartIndicator::UpdateIndicator(int32_t value) {
+void HorizontalChartIndicator::UpdateIndicator(float value) {
     auto ser = lv_chart_get_series_next(lv_chart_, nullptr);
 
-    lv_chart_set_next_value(lv_chart_, ser, value);
-    value_ = static_cast<float>(value);
+    lv_chart_set_next_value(lv_chart_, ser, static_cast<int32_t>(value));
+    value_ = value;
 }
 
 void HorizontalChartIndicator::Configure(const WidgetConfiguration& config) {
